follow change in type of ProcessHeader::m_randomSeeds
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@137159 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 43ba6ffe32fb4d9b40a06cf9f4920b10af420a2c

+++ LHCb/Event/EventPacker/src/ReadPackedDst.cpp
@@ -1,4 +1,3 @@
-// $Id: ReadPackedDst.cpp,v 1.13 2010-05-18 09:03:21 jonrob Exp $
 // Include files
 
 // from Gaudi
@@ -238,7 +237,7 @@ StatusCode ReadPackedDst::execute() {
       eventNumber = (eventNumber << 32) + nextInt();
       recHeader->setEvtNumber( eventNumber );
       unsigned nb = nextInt();
-      std::vector<long int> seeds;
+      std::vector<long long> seeds;
       for ( unsigned int kk=0 ; nb > kk; ++kk ) {
         seeds.push_back( nextInt() );
       }
